Vhardware/requirements.txt
Flask
pyserial
requests
